// src/api/model/apiPaginatedListApiBulkPayment.ts
export interface ApiPaginatedListApiBulkPayment { 
    /**
     * Response items.
     */
    items?: Array<ApiBulkPayment>;
    /**
     * Count of all items satisfying 'paginatable' request.
     */
    count?: number;
    /**
     * Offset got from request
     */
    offset?: number;
    /**
     * Limit got from request
     */
    limit?: number;
}

/**
 * Namespace for property- and property-value-enumerations of ApiPaginatedListApiBulkPayment.
 */
export namespace ApiPaginatedListApiBulkPayment {
    /**
     * All properties of ApiPaginatedListApiBulkPayment.
     */
    export enum Properties {
        /**
         * Response items.
         */
        items = 'items',
        /**
         * Count of all items satisfying 'paginatable' request.
         */
        count = 'count',
        /**
         * Offset got from request
         */
        offset = 'offset',
        /**
         * Limit got from request
         */
        limit = 'limit'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiPaginatedListApiBulkPayment',
            vars: [
                {
                    metadata: ApiBulkPayment.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'items',
                    classname: 'ApiPaginatedListApiBulkPayment',
                    dataType: 'Array&lt;ApiBulkPayment&gt;',
                    isPrimitiveType: false,
                    isListContainer: true,
                    complexType: 'ApiBulkPayment'
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'count',
                    classname: 'ApiPaginatedListApiBulkPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'offset',
                    classname: 'ApiPaginatedListApiBulkPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'limit',
                    classname: 'ApiPaginatedListApiBulkPayment',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                items: [
                ],
                count: [
                ],
                offset: [
                ],
                limit: [
                ],
            }
        }
    }

  // export const ApiPaginatedListApiBulkPaymentValidationScheme = {
  //     validators: [],
  //     fields: {
  //               items: {
  //                   validators: []
  //               },
  //               count: {
  //                   validators: []
  //               },
  //               offset: {
  //                   validators: []
  //               },
  //               limit: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiPaginatedListApiBulkPayment>;


}